Rename file from choices to game Game logic will be included in this file
