+++ app/javascript/packs/application.js
@@ -9,7 +9,7 @@ require("@rails/activestorage").start()
 require("channels")
 require('jquery')
 require('../stylesheets/application.scss')
-require('./choices')
+require('./game')
 import 'bootstrap/dist/js/bootstrap'
 import 'bootstrap/scss/bootstrap'
 import Playlist from './playlist'
